SB-0MN9CIUK5008VPR6: Extract Discord-free shared types
- Create src/interfaces/cli-types.ts with all progress/ingestion types
  - ProgressPhase, ProgressUpdate, IngestionProgress, ProgressCallback
  - CliQueueItem (Discord-free QueueItem variant)
  - QueueUpdateStatus, PendingQueueItem, QueueUpdateCallback
  - CrawlPhase, CrawlProgress, CrawlResult, CrawlProgressCallback
  - RecoveryResult, ShutdownSignal, ShutdownOptions
  - Result, AsyncInitializable, Disposable, Logger
  - SyntheticMessage

- Refactor src/interfaces/types.ts to re-export shared types
  - Keep Discord-specific types: MessageUrl, CommandContext, QueueItem
  - Re-export all CLI-compatible types for backward compatibility

- Create src/interfaces/cli-progress-presenter.ts
  - CliProgressPresenterOptions (Discord-free base)
  - CliPresenterResult, CliProgressPresenter, CliProgressPresenterFactory
  - CliProgressPresenterBase abstract class
  - PHASE_EMOJI and PHASE_LABEL constants

- Update src/interfaces/progress-presenter.ts
  - Extend CLI base interfaces with Discord-specific options
  - ProgressPresenterOptions now extends CliProgressPresenterOptions
  - PresenterResult extends CliPresenterResult
  - ProgressPresenterBase extends CliProgressPresenterBase

- Update src/interfaces/queue-presenter.ts
  - Import from cli-types.ts instead of types.ts
  - Reference CliQueueItem in docstring examples

- Update src/interfaces/index.ts
  - Export CLI types and CLI progress presenter interfaces
  - Organize exports by CLI-compatible vs Discord-specific

- TypeScript compilation succeeds with no errors
- No files under src/cli/ transitively import discord.js
